feat: get all products with pagination

+++ src/main/java/com/server/ecommerce/product/domain/Product.java
@@ -24,7 +24,7 @@ public class Product extends BaseEntity {
 
     private String description;
 
-    private int price;
+    private long price;
 
     @Column(name = "is_enabled")
     private boolean isEnabled;
