+++ src/main/java/com/server/ecommerce/product/dto/ProductCreateDTO.java
@@ -1,4 +1,4 @@
 package com.server.ecommerce.product.dto;
 
-public record ProductCreateDTO(String name, String description, int price, int quantity, boolean isEnabled){
+public record ProductCreateDTO(String name, String description, long price, int quantity, boolean isEnabled){
 }
